Gyorskereső szűrők
A gyorskeresőben beállíthatók a oszlopok szűrői,
amivel megkönnyíthető a keresés.
